+++ src/app/core/services/api/fake-backend/connector-fake-impl/asset-fake-service.ts
@@ -15,6 +15,7 @@ let assets: UiAsset[] = [
   TestAssets.onRequestAsset,
   TestAssets.boring,
   TestAssets.short,
+  TestAssets.assetWithCustomProperties,
 ];
 
 export const assetPage = (): AssetPage => {